sound_machine/requirements.txt
pyserial>=3.5
pyttsx3>=2.90
python-dotenv>=1.0.1
